add action in one file
