--- .eslintrc.json
@@ -1,44 +1,0 @@
-{
-    "env": {
-        "browser": true,
-        "es6": true,
-        "jest": true
-    },
-    "parser": "babel-eslint",
-    "parserOptions": {
-        "ecmaFeatures": {
-            "jsx": true
-        },
-        "ecmaVersion": 2018,
-        "sourceType": "module"
-    },
-    "extends": [
-        "airbnb",
-        "plugin:react/recommended"
-    ],
-    "plugins": [
-        "react"
-    ],
-    "rules": {
-        "react/jsx-filename-extension": [
-            "warn",
-            {
-                "extensions": [
-                    ".js",
-                    ".jsx"
-                ]
-            }
-        ],
-        "react/react-in-jsx-scope": "off",
-        "import/no-unresolved": "off",
-        "no-shadow": "off",
-        "arrow-parens": [
-            "error",
-            "as-needed"
-        ]
-    },
-    "ignorePatterns": [
-        "dist/",
-        "build/"
-    ]
-}
