--- src/actions/favouriteTypes.js
@@ -1,6 +1,0 @@
-export const FETCH_FAVOURITES_REQUEST = 'FETCH_FAVOURITES_REQUEST';
-export const FETCH_FAVOURITES_SUCCESS = 'FETCH_FAVOURITES_SUCCESS';
-export const FETCH_FAVOURITES_FAILURE = 'FETCH_FAVOURITES_FAILURE';
-export const CREATE_FAVOURITE_REQUEST = 'CREATE_FAVOURITE_REQUEST';
-export const CREATE_FAVOURITE_SUCCESS = 'CREATE_FAVOURITE_SUCCESS';
-export const CREATE_FAVOURITE_FAILURE = 'CREATE_FAVOURITE_FAILURE';
