+++ src/actions/index.js
@@ -10,6 +10,20 @@ import {
   CREATE_BOOKMOTO_FAILURE,
   CREATE_BOOKMOTO_REQUEST,
   CREATE_BOOKMOTO_SUCCESS,
+  FETCH_USERS_FAILURE,
+  FETCH_USERS_REQUEST,
+  FETCH_USERS_SUCCESS,
+  CREATE_USERS_FAILURE,
+  CREATE_USERS_REQUEST,
+  CREATE_USERS_SUCCESS,
+  LOGIN_USER,
+  LOGOUT_USER,
+  FETCH_FAVOURITES_FAILURE,
+  FETCH_FAVOURITES_REQUEST,
+  FETCH_FAVOURITES_SUCCESS,
+  CREATE_FAVOURITE_FAILURE,
+  CREATE_FAVOURITE_REQUEST,
+  CREATE_FAVOURITE_SUCCESS,
 
 } from './types';
 
@@ -55,6 +69,70 @@ const createBookMotoFailure = error => ({
   payload: error,
 });
 
+const fetchUsersRequest = () => ({
+  type: FETCH_USERS_REQUEST,
+});
+
+const fetchUsersSuccess = users => ({
+  type: FETCH_USERS_SUCCESS,
+  payload: users,
+});
+
+const fetchUsersFailure = error => ({
+  type: FETCH_USERS_FAILURE,
+  payload: error,
+});
+
+const createUsersRequest = () => ({
+  type: CREATE_USERS_REQUEST,
+});
+
+export const createUsersSuccess = user => ({
+  type: CREATE_USERS_SUCCESS,
+  payload: user,
+});
+
+const createUsersFailure = error => ({
+  type: CREATE_USERS_FAILURE,
+  payload: error,
+});
+
+export const loginUser = () => ({
+  type: LOGIN_USER,
+});
+
+export const logoutUser = () => ({
+  type: LOGOUT_USER,
+});
+
+const fetchFavouritesRequest = () => ({
+  type: FETCH_FAVOURITES_REQUEST,
+});
+
+const fetchFavouritesSuccess = motos => ({
+  type: FETCH_FAVOURITES_SUCCESS,
+  payload: motos,
+});
+
+const fetchFavouritesFailure = error => ({
+  type: FETCH_FAVOURITES_FAILURE,
+  payload: error,
+});
+
+const createFavouriteRequest = () => ({
+  type: CREATE_FAVOURITE_REQUEST,
+});
+
+const createFavouriteSuccess = moto => ({
+  type: CREATE_FAVOURITE_SUCCESS,
+  payload: moto,
+});
+
+const createFavouriteFailure = error => ({
+  type: CREATE_FAVOURITE_FAILURE,
+  payload: error,
+});
+
 export const fetchMotos = () => dispatch => {
   dispatch(fetchMotosRequest());
   axios.get('/motocycles')
@@ -96,5 +174,87 @@ export const createMotoBook = (motoid, date, city) => dispatch => {
     })
     .catch(error => {
       dispatch(createBookMotoFailure(error.message));
+    });
+};
+
+export const fetchUsers = (email, password) => dispatch => {
+  dispatch(fetchUsersRequest);
+
+  axios.post('/login', {
+
+    email,
+    password,
+  })
+    .then(response => {
+      const user = response.data;
+
+      localStorage.setItem('motoToken', user.token);
+
+      if (user) {
+        dispatch(fetchUsersSuccess(user));
+        dispatch(loginUser());
+      } else {
+        dispatch(fetchUsersFailure('user do not exist'));
+      }
+    })
+    .catch(error => {
+      dispatch(fetchUsersFailure(error.message));
+    });
+};
+
+export const createUsers = (name, email, password) => dispatch => {
+  dispatch(createUsersRequest);
+
+  axios.post('/register',
+    {
+      name,
+      email,
+      password,
+    })
+    .then(response => {
+      const user = response.data;
+      localStorage.setItem('motoToken', user.token);
+      if (user) {
+        dispatch(createUsersSuccess(response.data));
+        dispatch(loginUser());
+      } else {
+        dispatch(createUsersFailure('This email is in use, write another email!'));
+      }
+    })
+    .catch(error => {
+      dispatch(createUsersFailure(error.message));
+    });
+};
+
+export const setUser = user => dispatch => {
+  dispatch(fetchUsersSuccess(user));
+};
+
+export const fetchFavourites = () => dispatch => {
+  dispatch(fetchFavouritesRequest);
+
+  axios.get('/favourites', { mode: 'cors' })
+    .then(response => {
+      const motos = response.data;
+      dispatch(fetchFavouritesSuccess(motos));
+    })
+    .catch(error => {
+      dispatch(fetchFavouritesFailure(error.message));
+    });
+};
+
+export const createFavourite = motoid => dispatch => {
+  dispatch(createFavouriteRequest);
+  axios.post('/favourites',
+    {
+      motocycle_id: motoid,
+    })
+    .then(response => {
+      const motosf = response.data;
+      console.log('all fine')
+      dispatch(createFavouriteSuccess(motosf));
+    })
+    .catch(error => {
+      dispatch(createFavouriteFailure(error.message));
     });
 };
