+++ src/actions/types.js
@@ -7,3 +7,17 @@ export const FETCH_BOOKMOTO_FAILURE = 'FETCH_BOOKMOTO_FAILURE';
 export const CREATE_BOOKMOTO_REQUEST = 'CREATE_BOOKMOTO_REQUEST';
 export const CREATE_BOOKMOTO_SUCCESS = 'CREATE_BOOKMOTO_SUCCESS';
 export const CREATE_BOOKMOTO_FAILURE = 'CREATE_BOOKMOTO_FAILURE';
+export const FETCH_USERS_REQUEST = 'FETCH_USERS_REQUEST';
+export const FETCH_USERS_SUCCESS = 'FETCH_USERS_SUCCESS';
+export const FETCH_USERS_FAILURE = 'FETCH_USERS_FAILURE';
+export const CREATE_USERS_REQUEST = 'CREATE_USERS_REQUEST';
+export const CREATE_USERS_SUCCESS = 'CREATE_USERS_SUCCESS';
+export const CREATE_USERS_FAILURE = 'CREATE_USERS_FAILURE';
+export const LOGIN_USER = 'LOGIN_USER';
+export const LOGOUT_USER = 'LOGOUT_USER';
+export const FETCH_FAVOURITES_REQUEST = 'FETCH_FAVOURITES_REQUEST';
+export const FETCH_FAVOURITES_SUCCESS = 'FETCH_FAVOURITES_SUCCESS';
+export const FETCH_FAVOURITES_FAILURE = 'FETCH_FAVOURITES_FAILURE';
+export const CREATE_FAVOURITE_REQUEST = 'CREATE_FAVOURITE_REQUEST';
+export const CREATE_FAVOURITE_SUCCESS = 'CREATE_FAVOURITE_SUCCESS';
+export const CREATE_FAVOURITE_FAILURE = 'CREATE_FAVOURITE_FAILURE';
