--- src/actions/userTypes.js
@@ -1,8 +1,0 @@
-export const FETCH_USERS_REQUEST = 'FETCH_USERS_REQUEST';
-export const FETCH_USERS_SUCCESS = 'FETCH_USERS_SUCCESS';
-export const FETCH_USERS_FAILURE = 'FETCH_USERS_FAILURE';
-export const CREATE_USERS_REQUEST = 'CREATE_USERS_REQUEST';
-export const CREATE_USERS_SUCCESS = 'CREATE_USERS_SUCCESS';
-export const CREATE_USERS_FAILURE = 'CREATE_USERS_FAILURE';
-export const LOGIN_USER = 'LOGIN_USER';
-export const LOGOUT_USER = 'LOGOUT_USER';
