+++ src/components/AsideMenu.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import {
   logoutUser,
-} from '../actions/userActions';
+} from '../actions/index';
 import './AsideMenu.css';
 
 const AsideMenu = ({ logoutUser }) => {
